refactor(graphql): improve the getQueryResults
LIIKUNTA-578
