+++ graphql/src/datasources/es/api/getQueryResults/getQueryResults.ts
@@ -2,32 +2,15 @@ import { DEFAULT_ELASTIC_LANGUAGE } from '../../../../types';
 import { ES_DEFAULT_INDEX } from '../../constants';
 import { type ElasticSearchAPI } from '../..';
 import type { getQueryResultsProps } from './types';
-import {
-  getDefaultQuery,
-  getOntologyFields,
-  getFilters,
-  sortQuery,
-} from './utils';
+import { getDefaultQuery, getOntologyFields, sortQuery } from './utils';
+import { filterQuery } from './utils/filterQuery';
 
-function getQuery({
+function createQuery({
   index,
   languages,
   text,
   ontology,
-  from,
-  size,
-  administrativeDivisionIds,
-  ontologyTreeIdOrSets,
-  ontologyWordIdOrSets,
-  providerTypes,
-  serviceOwnerTypes,
-  targetGroups,
-  mustHaveReservableResource,
-  openAt,
-  orderByAccessibilityProfile,
-  orderByDistance,
-  orderByName,
-}: getQueryResultsProps) {
+}: Pick<getQueryResultsProps, 'index' | 'languages' | 'text' | 'ontology'>) {
   const defaultQuery = getDefaultQuery({ index, languages, text });
 
   // Resolve query
@@ -66,8 +49,31 @@ function getQuery({
       },
     };
   }
+  return query;
+}
 
-  const filters = getFilters({
+function queryBuilder({
+  index,
+  languages,
+  text,
+  ontology,
+  from,
+  size,
+  administrativeDivisionIds,
+  ontologyTreeIdOrSets,
+  ontologyWordIdOrSets,
+  providerTypes,
+  serviceOwnerTypes,
+  targetGroups,
+  mustHaveReservableResource,
+  openAt,
+  orderByAccessibilityProfile,
+  orderByDistance,
+  orderByName,
+}: getQueryResultsProps) {
+  const query = createQuery({ index, languages, text, ontology });
+
+  filterQuery(query, {
     administrativeDivisionIds,
     ontologyTreeIdOrSets,
     ontologyWordIdOrSets,
@@ -77,11 +83,6 @@ function getQuery({
     mustHaveReservableResource,
     openAt,
   });
-
-  if (filters.length > 0) {
-    query.query.bool.minimum_should_match = 1;
-    query.query.bool.filter = filters;
-  }
 
   const language = languages[0] ?? DEFAULT_ELASTIC_LANGUAGE;
 
@@ -103,7 +104,7 @@ export default async function getQueryResults(
   request: ElasticSearchAPI['post'],
   { index = ES_DEFAULT_INDEX, ...queryProps }: getQueryResultsProps
 ) {
-  const query = getQuery({ index, ...queryProps });
+  const query = queryBuilder({ index, ...queryProps });
 
   return await request(`${index}/_search`, undefined, {
     headers: { 'Content-Type': 'application/json' },
